examples/microservices/data_generator/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.6
Faker==18.9.0
